1. nominal yang harus dibayar pembeli   2. tabel *daftar kendaraan* ✅ 3. user input kendaraan  4. harga pakai tipe long ✅

+++ showroom kendaraan/src/Vehicle.java
@@ -1,11 +1,11 @@
 public class Vehicle {
     private String brand;
     private String model;
-    private int harga;
+    private long harga;
     private int stockQuantity;
 
     // Constructor
-    public Vehicle(String brand, String model, int harga, int stockQuantity) {
+    public Vehicle(String brand, String model, long harga, int stockQuantity) {
         this.brand = brand;
         this.model = model;
         this.harga = harga;
@@ -29,11 +29,11 @@ public class Vehicle {
         this.model = model;
     }
 
-    public int getHarga() {
+    public long getHarga() {
         return harga;
     }
 
-    public void setHarga(int harga) {
+    public void setHarga(long harga) {
         this.harga = harga;
     }
